+++ projects/s7-midi/midi_module.c
@@ -1083,6 +1083,8 @@ static s7_pointer g_write_mid(s7_scheme *sc, s7_pointer args) {
             case 2:  /* cc */
                 midi_file_writer_cc(writer, tick, 0, channel, e->data1, e->data2);
                 break;
+            default:
+                break;
         }
     }
 
@@ -1121,6 +1123,7 @@ static void read_mid_callback(void* ctx, const midi_file_event* event) {
         case 0xC0: type_name = "program"; break;
         case 0xD0: type_name = "aftertouch"; break;
         case 0xE0: type_name = "pitch-bend"; break;
+        default: break;
     }
 
     /* Create event as alist */